requirements.txt
click>=8.0.0
feedparser>=6.0.10
requests>=2.31.0
pyyaml>=6.0
rich>=13.0.0
mutagen>=1.47.0

